refactored some classes and functions
