+++ app/src/main/java/ir/shecan/dialog/RenewalDialog.java
@@ -1,0 +1,53 @@
+package ir.shecan.dialog;
+
+import android.app.Activity;
+import android.app.AlertDialog;
+import android.content.Context;
+import android.content.Intent;
+import android.net.Uri;
+import android.os.Build;
+import android.view.View;
+import android.view.WindowManager;
+import android.widget.Button;
+
+import androidx.annotation.RequiresApi;
+
+import ir.shecan.R;
+import ir.shecan.Shecan;
+
+public class RenewalDialog {
+    final Activity activity;
+
+    public RenewalDialog(Activity activity) {
+        this.activity = activity;
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.O)
+    public void show() {
+        View dialogView = activity.getLayoutInflater().inflate(R.layout.dialog_no_active_service, null);
+
+        AlertDialog.Builder builder = new AlertDialog.Builder(activity);
+        builder.setView(dialogView);
+
+        Button openUrlButton = dialogView.findViewById(R.id.renewalButton);
+        final AlertDialog dialog = builder.create();
+
+        if (dialog.getWindow() != null) {
+            dialog.getWindow().setLayout(WindowManager.LayoutParams.MATCH_PARENT, WindowManager.LayoutParams.MATCH_PARENT);
+            dialog.getWindow().setBackgroundDrawableResource(android.R.color.transparent);
+        }
+
+        openUrlButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                dialog.dismiss();
+
+                String url = Shecan.ShecanInfo.getPurchaseLink(); // Replace with your desired URL
+                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
+                activity.startActivity(intent);
+            }
+        });
+
+        dialog.show();
+    }
+}
